+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.70.0-sumo-0]
+## [v0.70.0-sumo-1]
 
 ### Added
 
@@ -15,12 +15,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(telegrafreceiver): make shutdown safe to call before start [#913]
 - chore: update OT core to v0.70.0 [#915]
 
-[v0.70.0-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.69.0-sumo-0...v0.70.0-sumo-0
+[v0.70.0-sumo-1]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.69.0-sumo-0...v0.70.0-sumo-1
 [#893]: https://github.com/SumoLogic/sumologic-otel-collector/pull/893
 [#885]: https://github.com/SumoLogic/sumologic-otel-collector/pull/885
 [#902]: https://github.com/SumoLogic/sumologic-otel-collector/pull/902
 [#913]: https://github.com/SumoLogic/sumologic-otel-collector/pull/913
 [#915]: https://github.com/SumoLogic/sumologic-otel-collector/pull/915
+
+## [v0.70.0-sumo-0]
+
+This was a failed release. Use [v0.70.0-sumo-1].
 
 ## [v0.69.0-sumo-0]